ML-models/caremel/second_layer/src/requirements.txt
python-dotenv==0.20.0
pandas==1.4.4
psycopg2-binary==2.9.3
numpy==1.23.2
scikit-learn==1.1.2
